+++ backend/mora/service/facet.py
@@ -708,14 +708,14 @@ async def get_all_classes_children(facet: str):
 
     facetids = await get_facetids(facet)
 
-    classes = await c.klasse.paged_get(
+    classes = (await c.klasse.paged_get(
         get_one_class,
         facet=facetids,
         ansvarlig=None,
         publiceret='Publiceret',
         start=start, limit=limit
-    )['items']
+    ))['items']
     classes = await gather(
-        *[create_task(prepare_class_child(c, class_) for class_ in classes)])
+        *[create_task(prepare_class_child(c, class_)) for class_ in classes])
 
     return flask.jsonify(classes)